Create Implementation of Authorization and Access Control.js
This code defines the different roles that users can have in Variable, the different resources that can be accessed in Variable, and the authorization rules for it. It also includes a function to check if a user has permission to access a resource and a function to authorize access to a resource.