Create Additional Secutity Measure - 2FA.js
This code checks if the user has logged in more than 5 times in a row. If the user has, the code throws an error. Otherwise, the code increments the user's login attempts and checks if the user is locked out. If the user is not locked out, the code logs the user in.