Create Encryption of Public addresses.js
code uses the crypto library to encrypt and decrypt public addresses. The crypto library is a secure and easy-to-use encryption library that is available in Node.JS.

The code first generates a random encryption key. This key should be long and complex, and it should be kept secret. The key is then used to encrypt the public address. The encrypted public address is then stored in the user's account database.

When the user logs in, the DEX can decrypt the public address using the same encryption key. This will allow the DEX to access the user's funds.

The code also implements some additional security measures to protect the user's funds. These measures include two-factor authentication and cold storage.