orderbook.js
To make an order book out of this code, you will need to do the following:

Create a new file and save it as orderbook.js.
Copy and paste the code from the previous response into the orderbook.js file.
Create a new directory and save it as orderbook-app.
In the orderbook-app directory, create a new file and save it as index.js.
In the index.js file, import the OrderBook class from the orderbook.js file.
Create a new instance of the OrderBook class and pass in the symbol of the asset that you want to track.
Use the addOrder method to add orders to the order book.
Use the removeOrder method to remove orders from the order book.
Use the getBestBid method to get the best bid price for the asset.
Use the getBestAsk method to get the best ask price for the asset.
Use the getOrders method to get all of the orders for the asset.
Once you have completed these steps, you will have a working order book. You can use this order book to track the prices of assets and to place orders.