Create balances.js
The below code will create a collection in the database called balances. The collection will store documents that represent the balance of each wallet for each cryptocurrency. The documents will also store a list of open positions and a list of closed positions for each wallet.

The cryptocurrencies that are supported are:

ETH
BTC
USDC
USDT
MATIC
SOL
To update the balance of a wallet, you can use the updateOne() method. The updateOne() method takes two arguments: the first argument is the filter that is used to find the document that you want to update, and the second argument is the update object. The update object specifies the changes that you want to make to the document.

To query the balance of a wallet, you can use the findOne() method. The findOne() method takes one argument: the filter that is used to find the document that you want to query.