Create mongoDB-Orderbook-matching-engine-connection.js
placeholders that can be easily filled in by other developers. We'll modify the existing connectToMongoDB function in the OrderMatchingEngine class to accept the necessary MongoDB connection parameters as arguments.

Developers can then provide the appropriate values for databaseURL, dbName, and collectionName when calling the connectToMongoDB function, enabling them to establish a connection to their specific MongoDB database and collection.