orderStatus.js
defines a class called OrderStatus that has methods for updating the status of an order. To use this code in your order matching engine, you will need to:

Import the OrderStatus class into your code.
Create an instance of the OrderStatus class for each order.
Use the methods on the OrderStatus class to update the status of the order.
Here is an example of how you could use the OrderStatus class in your order matching engine:

Code snippet
import OrderStatus from './OrderStatus';

// Create an instance of the OrderStatus class for each order.
const order1 = new OrderStatus();
const order2 = new OrderStatus();

// Update the status of the order.
order1.markOpen();
order2.markFilled();
Use code with caution. Learn more
The OrderStatus class provides a simple way to track the status of orders in your order matching engine. You can use the methods on this class to update the status of orders as they are placed, filled, cancelled, or modified.

Here is a table that summarizes the methods on the OrderStatus class:

Method	Description
constructor()	Creates a new instance of the OrderStatus class.
markOpen()	Updates the status of the order to Open.
markPartiallyFilled()	Updates the status of the order to Partially Filled.
markFilled()	Updates the status of the order to Filled.
markCancelled()	Updates the status of the order to Cancelled.
markModified()	Updates the status of the order to Modified.