Create userBalances.js
Sure. Here is an example of how to use the `updateBalance()` function:

```
const UserBalances = require('./UserBalances');

// Update the balance of the user with the ID `user1` to 10 BTC
UserBalances.updateBalance('user1', 'BTC', 10);
```

This code will call the `updateBalance()` function from the `UserBalances` library and update the balance of the user with the ID `user1` to 10 BTC.

Here is another example of how to use the `getBalance()` function in JavaScript (Node.js):

```
const UserBalances = require('./UserBalances');

// Get the balance of the user with the ID `user1`
const balance = UserBalances.getBalance('user1', 'BTC');

console.log(`User ${user1} balance for ${asset}: ${balance}`);
```

This code will call the `getBalance()` function from the `UserBalances` library and print the balance of the user with the ID `user1` to the console.